+++ angular_analyzer_plugin/lib/errors.dart
@@ -26,6 +26,8 @@ const _angularWarningCodeValues = const <AngularWarningCode>[
   AngularWarningCode.NONEXIST_TWO_WAY_OUTPUT_BOUND,
   AngularWarningCode.TWO_WAY_BINDING_OUTPUT_TYPE_ERROR,
   AngularWarningCode.INPUT_BINDING_TYPE_ERROR,
+  AngularWarningCode.ATTR_IF_BINDING_TYPE_ERROR,
+  AngularWarningCode.UNMATCHED_ATTR_IF_BINDING,
   AngularWarningCode.TRAILING_EXPRESSION,
   AngularWarningCode.OUTPUT_MUST_BE_STREAM,
   AngularWarningCode.TWO_WAY_BINDING_NOT_ASSIGNABLE,
@@ -62,6 +64,8 @@ const _angularWarningCodeValues = const <AngularWarningCode>[
   AngularWarningCode.PIPE_REQUIRES_TRANSFORM_METHOD,
   AngularWarningCode.PIPE_TRANSFORM_NO_NAMED_ARGS,
   AngularWarningCode.PIPE_TRANSFORM_REQ_ONE_ARG,
+  AngularWarningCode.PIPE_NOT_FOUND,
+  AngularWarningCode.AMBIGUOUS_PIPE,
   AngularWarningCode.UNSAFE_BINDING,
   AngularWarningCode.EVENT_REDUCTION_NOT_ALLOWED,
   AngularWarningCode.FUNCTIONAL_DIRECTIVES_CANT_BE_EXPORTED,
@@ -227,6 +231,19 @@ class AngularWarningCode extends ErrorCode {
       'INPUT_BINDING_TYPE_ERROR',
       'Attribute value expression (of type {0}) is not assignable to component'
       ' input (of type {1})');
+
+  /// An error code indicating that an [attr.foo.if] binding was bound to an
+  /// expression that was not of type bool
+  static const ATTR_IF_BINDING_TYPE_ERROR = const AngularWarningCode(
+      'ATTR_IF_BINDING_TYPE_ERROR',
+      'Attribute value expression (of type {0}) must be of type bool');
+
+  /// An error code indicating that an [attr.foo.if] binding was used, but no
+  /// [attr.foo] binding simultaneously occured.
+  static const UNMATCHED_ATTR_IF_BINDING = const AngularWarningCode(
+      'UNMATCHED_ATTR_IF_BINDING',
+      'attr-if binding for attr {0} does not have a corresponding'
+      ' [attr.{0}] binding for it to affect');
 
   /// An error code indicating that an expression did not correctly
   /// end with an EOF token.
@@ -474,6 +491,20 @@ class AngularWarningCode extends ErrorCode {
   static const PIPE_TRANSFORM_REQ_ONE_ARG = const AngularWarningCode(
       'PIPE_TRANSFORM_REQ_ONE_ARG',
       "'transform' method requires at least one argument");
+
+  /// An error indicating that pipe syntax was used in an angular template, but
+  /// the name of the pipe doesn't match one defined in the component
+  static const PIPE_NOT_FOUND = const AngularWarningCode(
+      'PIPE_NOT_FOUND',
+      "Pipe by name of {0} not found. Did you reference it in your @Component"
+      " configuration?");
+
+  /// An error indicating that pipe syntax was used in an angular template, but
+  /// the name of the pipe doesn't match one defined in the component
+  static const AMBIGUOUS_PIPE = const AngularWarningCode(
+      'AMBIGUOUS_PIPE',
+      "Multiple pipes by name of {0} found. Check the `pipes` field of your "
+      "@Component annotation for duplicates and/or conflicts.");
 
   /// An error indicating that a security exception will be thrown by this input
   /// binding