+++ angular_analyzer_plugin/lib/src/resolver.dart
@@ -1,5 +1,3 @@
-library angular2.src.analysis.analyzer_plugin.src.resolver;
-
 import 'dart:collection';
 
 import 'package:analyzer/dart/ast/ast.dart' hide Directive;
@@ -10,6 +8,7 @@ import 'package:analyzer/dart/element/type.dart';
 import 'package:analyzer/error/error.dart';
 import 'package:analyzer/error/listener.dart';
 import 'package:analyzer/src/dart/element/element.dart';
+import 'package:analyzer/src/error/codes.dart';
 import 'package:analyzer/src/generated/error_verifier.dart';
 import 'package:analyzer/src/generated/resolver.dart';
 import 'package:analyzer/src/generated/source.dart';
@@ -107,10 +106,11 @@ class AngularErrorVerifier extends _IntermediateErrorVerifier
 class AngularResolverVisitor extends _IntermediateResolverVisitor
     with ReportUnacceptableNodesMixin {
   final bool acceptAssignment;
+  final List<Pipe> pipes;
 
   AngularResolverVisitor(LibraryElement library, Source source,
       TypeProvider typeProvider, AnalysisErrorListener errorListener,
-      {@required this.acceptAssignment})
+      {@required this.acceptAssignment, @required this.pipes})
       : super(library, source, typeProvider, errorListener);
 
   @override
@@ -118,6 +118,27 @@ class AngularResolverVisitor extends _IntermediateResolverVisitor
     // This means we generated this in a pipe, and its OK.
     if (exp.asOperator.offset == 0) {
       super.visitAsExpression(exp);
+      final pipeName = exp.getProperty<SimpleIdentifier>('_ng_pipeName');
+      final matchingPipes =
+          pipes.where((pipe) => pipe.pipeName == pipeName.name);
+      if (matchingPipes.isEmpty) {
+        errorReporter.reportErrorForNode(
+            AngularWarningCode.PIPE_NOT_FOUND, pipeName, [pipeName]);
+      } else if (matchingPipes.length > 1) {
+        errorReporter.reportErrorForNode(
+            AngularWarningCode.AMBIGUOUS_PIPE, pipeName, [pipeName]);
+      } else {
+        final matchingPipe = matchingPipes.single;
+        exp.staticType = matchingPipe.transformReturnType;
+
+        if (!typeSystem.isAssignableTo(
+            exp.expression.staticType, matchingPipe.requiredArgumentType)) {
+          errorReporter.reportErrorForNode(
+              StaticWarningCode.ARGUMENT_TYPE_NOT_ASSIGNABLE,
+              exp.expression,
+              [exp.expression.staticType, matchingPipe.requiredArgumentType]);
+        }
+      }
     } else {
       _reportUnacceptableNode(exp, "As expression");
     }
@@ -352,7 +373,9 @@ class ComponentContentResolver extends AngularAstVisitor {
   bool _isStandardTagName(String name) {
     // ignore: parameter_assignments
     name = name.toLowerCase();
-    return !name.contains('-') || name == 'ng-content';
+    return !name.contains('-') ||
+        name == 'ng-content' ||
+        name == 'ng-container';
   }
 
   void _reportErrorForRange(SourceRange range, ErrorCode errorCode,
@@ -1207,6 +1230,7 @@ class SingleScopeResolver extends AngularScopeVisitor {
   static final RegExp _cssIdentifierRegexp =
       new RegExp(r"^(-?[a-zA-Z_]|\\.)([a-zA-Z0-9\-_]|\\.)*$");
   final Map<String, InputElement> standardHtmlAttributes;
+  final List<Pipe> pipes;
   List<AbstractDirective> directives;
   View view;
   Template template;
@@ -1223,6 +1247,7 @@ class SingleScopeResolver extends AngularScopeVisitor {
 
   SingleScopeResolver(
       this.standardHtmlAttributes,
+      this.pipes,
       this.view,
       this.template,
       this.templateSource,
@@ -1280,6 +1305,8 @@ class SingleScopeResolver extends AngularScopeVisitor {
       _resolveStyleAttribute(attribute);
     } else if (attribute.bound == ExpressionBoundType.attr) {
       _resolveAttributeBoundAttribute(attribute);
+    } else if (attribute.bound == ExpressionBoundType.attrIf) {
+      _resolveAttributeBoundAttributeIf(attribute);
     }
   }
 
@@ -1428,6 +1455,55 @@ class SingleScopeResolver extends AngularScopeVisitor {
     // failing tests for this)
   }
 
+  /// Resolve attributes of type [attribute.some-attribute]="someExpr"
+  void _resolveAttributeBoundAttributeIf(ExpressionBoundAttribute attribute) {
+    if (attribute.parent is! ElementInfo) {
+      assert(false, 'Got an attr-if bound attribute on non element! Aborting!');
+      return;
+    }
+
+    final parent = attribute.parent as ElementInfo;
+
+    // For the [attr.foo.if] attribute, find the matching [attr.foo] attribute.
+    final matchingAttr = parent.attributes
+        .where((attr) =>
+            attr is ExpressionBoundAttribute &&
+            attr.bound == ExpressionBoundType.attr)
+        .firstWhere((attrAttr) => attrAttr.name == attribute.name,
+            orElse: () => null);
+
+    // Error: no matching attribute to make conditional via this attr-if.
+    if (matchingAttr == null) {
+      errorListener.onError(new AnalysisError(
+          templateSource,
+          attribute.nameOffset,
+          attribute.name.length,
+          AngularWarningCode.UNMATCHED_ATTR_IF_BINDING,
+          [attribute.name]));
+      return;
+    }
+
+    // Add navigation from [attribute] (`[attr.foo.if]`) to [matchingAttr]
+    // (`[attr.foo]`).
+    final range = new SourceRange(attribute.nameOffset, attribute.name.length);
+    template.addRange(
+        range,
+        new AngularElementImpl('attr.${attribute.name}',
+            matchingAttr.nameOffset, matchingAttr.name.length, templateSource));
+
+    // Ensure the if condition was a boolean.
+    if (attribute.expression != null &&
+        !typeSystem.isAssignableTo(
+            attribute.expression.staticType, typeProvider.boolType)) {
+      errorListener.onError(new AnalysisError(
+          templateSource,
+          attribute.valueOffset,
+          attribute.value.length,
+          AngularWarningCode.ATTR_IF_BINDING_TYPE_ERROR,
+          [attribute.name]));
+    }
+  }
+
   /// Resolve attributes of type [class.some-class]="someBoolExpr", ensuring
   /// the class is a valid css identifier and that the expression is of boolean
   /// type
@@ -1466,7 +1542,7 @@ class SingleScopeResolver extends AngularScopeVisitor {
     }
     final resolver = new AngularResolverVisitor(
         library, templateSource, typeProvider, errorListener,
-        acceptAssignment: acceptAssignment);
+        acceptAssignment: acceptAssignment, pipes: pipes);
     // fill the name scope
     final classScope = new ClassScope(resolver.nameScope, classElement);
     final localScope = new EnclosedScope(classScope);
@@ -1787,6 +1863,7 @@ class TemplateResolver {
         // Resolve the scopes
         ..accept(new SingleScopeResolver(
             standardHtmlAttributes,
+            view.pipes,
             view,
             template,
             templateSource,